1-3107: move stale flags to a seprate component + fix scaling issues in health
